feat: add error handling types and components for exam document management
