+++ src/components/AssessmentInfoCard/index.ts
@@ -1,0 +1,1 @@
+export { AssessmentInfoCard } from './AssessmentInfoCard';
