+++ src/components/LecturerFilesSection/index.ts
@@ -1,0 +1,1 @@
+export { LecturerFilesSection } from './LecturerFilesSection';
