--- src/components/Modals/ExamDocumentModal/index.ts
@@ -1,6 +1,0 @@
-export { default } from './ExamDocumentModal';
-export { UploadSection } from './UploadSection';
-export { StudentDocumentsList } from './StudentDocumentsList';
-export { LecturerFilesSection } from './LecturerFilesSection';
-export { ErrorBanner } from './ErrorBanner';
-export { AssessmentInfoCard } from './AssessmentInfoCard';
