+++ src/components/StudentDocumentsList/index.ts
@@ -1,0 +1,1 @@
+export { StudentDocumentsList } from './StudentDocumentsList';
