+++ src/components/UploadSection/index.ts
@@ -1,0 +1,1 @@
+export { UploadSection } from './UploadSection';
